+++ src/hooks/offer.ts
@@ -98,7 +98,7 @@ export function useMakeOffer(nft?: Nft): MakeOfferContext {
       );
     } else if (nft?.moonrankCollection?.trends?.compactFloor1d) {
       const minOffer =
-        parseInt(nft?.moonrankCollection?.trends?.compactFloor1d) *
+        parseFloat(nft?.moonrankCollection?.trends?.compactFloor1d) *
         config.offerMinimums.percentageFloor;
 
       validation = validation.min(
@@ -321,7 +321,7 @@ export function useUpdateOffer(offer: Maybe<Offer> | undefined, nft?: Nft): Upda
       );
     } else if (nft?.moonrankCollection?.trends?.compactFloor1d) {
       const minOffer =
-        parseInt(nft?.moonrankCollection?.trends?.compactFloor1d) *
+        parseFloat(nft?.moonrankCollection?.trends?.compactFloor1d) *
         config.offerMinimums.percentageFloor;
 
       validation = validation.min(
